Fix populate test data cmd
