+++ integration/src/main/resources/avro-integration-test/folder-a/folder-b/shakespeare-schema.avsc
