+++ integration/src/main/resources/shakespeare-sample-10.avro
