+++ integration/src/main/scala/com/spotify/scio/jdbc/PopulateTestData.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Spotify AB.
+ * Copyright 2024 Spotify AB.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,56 +15,23 @@
  * under the License.
  */
 
-package com.spotify.scio
+package com.spotify.scio.jdbc
 
-import org.apache.beam.sdk.io.FileSystems
-import org.apache.beam.sdk.options.PipelineOptionsFactory
-import org.apache.beam.sdk.util.MimeTypes
-import org.slf4j.LoggerFactory
-
-import java.nio.channels.Channels
-import java.nio.file.{Files, Path, Paths}
-import java.util.stream.Collectors
-import scala.jdk.CollectionConverters._
+import com.spotify.scio.jdbc.sharded.JdbcUtils
 
 object PopulateTestData {
-  private lazy val log = LoggerFactory.getLogger(getClass)
-
-  def main(args: Array[String]): Unit = {
-    populateFiles("data-integration-test-eu")
-    populateSql()
-  }
-
-  def populateFiles(bucket: String): Unit = {
-    FileSystems.setDefaultPipelineOptions(PipelineOptionsFactory.create())
-
-    val root = Paths.get("src/test/resources")
-    Files
-      .walk(root)
-      .collect(Collectors.toList[Path])
-      .asScala
-      .filter(Files.isRegularFile(_))
-      .foreach { src =>
-        val resourceId =
-          FileSystems.matchNewResource(s"gs://$bucket/${root.relativize(src)}", false)
-        val dst = Channels.newOutputStream(FileSystems.create(resourceId, MimeTypes.BINARY))
-        Files.copy(src, dst)
-        dst.close()
-        log.info(s"Populated file $resourceId.")
-      }
-  }
+  def main(args: Array[String]): Unit =
+    populate()
 
   // See https://learn.microsoft.com/en-us/sql/connect/ado-net/sql/compare-guid-uniqueidentifier-values?view=sql-server-ver16
-  def populateSql(): Unit = {
-    import com.spotify.scio.jdbc.sharded.JdbcUtils
-    import com.spotify.scio.jdbc.JdbcIOIT._
+  def populate(): Unit = {
 
-    val conn = JdbcUtils.createConnection(connection)
+    val conn = JdbcUtils.createConnection(SqlServer.connection)
     try {
       val stmt = conn.createStatement()
       val query =
-        s"""DROP TABLE IF EXISTS $tableId;
-           |CREATE TABLE $tableId
+        s"""DROP TABLE IF EXISTS employee;
+           |CREATE TABLE employee
            |(
            |    guid UNIQUEIDENTIFIER
            |        CONSTRAINT guid_default DEFAULT
@@ -73,7 +40,7 @@ object PopulateTestData {
            |
            |    CONSTRAINT guid_pk PRIMARY KEY (guid)
            |);
-           |INSERT INTO $tableId (guid, name)
+           |INSERT INTO employee (guid, name)
            |VALUES
            | (CAST('3AAAAAAA-BBBB-CCCC-DDDD-2EEEEEEEEEEE' AS UNIQUEIDENTIFIER), 'Bob'),
            | (CAST('2AAAAAAA-BBBB-CCCC-DDDD-1EEEEEEEEEEE' AS UNIQUEIDENTIFIER), 'Alice'),
