+++ integration/src/main/scala/com/spotify/scio/jdbc/SqlServer.scala
@@ -1,0 +1,21 @@
+package com.spotify.scio.jdbc
+
+object SqlServer {
+  val projectId = "data-integration-test"
+  val regionId = "us-central1"
+  val sqlServerInstanceId = "scio-sql-server-it"
+  val databaseId = "shard-it"
+  val username = "sqlserver"
+  val password = sys.props.get("cloudsql.sqlserver.password")
+
+  val connection = JdbcConnectionOptions(
+    username,
+    password,
+    s"jdbc:sqlserver://localhost;" +
+      "socketFactoryClass=com.google.cloud.sql.sqlserver.SocketFactory;" +
+      s"socketFactoryConstructorArg=$projectId:$regionId:$sqlServerInstanceId;" +
+      s"databaseName=$databaseId;" +
+      "encrypt=false", // otherwise we'll have to generate certificates
+    classOf[com.microsoft.sqlserver.jdbc.SQLServerDriver]
+  )
+}
