+++ integration/src/test/scala/com/spotify/scio/bigquery/types/BigQueryTypeIT.scala
@@ -87,7 +87,7 @@ object BigQueryTypeIT {
   class ShakespeareWithSequentialAnnotations
 }
 
-// scio-test/it:runMain com.spotify.scio.PopulateTestData to re-populate data for integration tests
+// integration/runMain com.spotify.scio.PopulateTestData to re-populate data for integration tests
 class BigQueryTypeIT extends AnyFlatSpec with Matchers {
   import BigQueryTypeIT._
